// flex-admin/src/main/webapp/app/content/auctions/enums/auction-status.ts
export enum AuctionStatus {
  NEW = 'NEW',
  OPEN = 'OPEN',
  CLOSED = 'CLOSED',

  NEW_CAPACITY = 'NEW_CAPACITY',
  OPEN_CAPACITY = 'OPEN_CAPACITY',
  CLOSED_CAPACITY = 'CLOSED_CAPACITY',
  NEW_ENERGY = 'NEW_ENERGY',
  OPEN_ENERGY = 'OPEN_ENERGY',
  CLOSED_ENERGY = 'CLOSED_ENERGY',
}

export enum AuctionStatusSimplified {
  OPEN_CAPACITY = 'OPEN_CAPACITY',
  CLOSED_CAPACITY = 'CLOSED_CAPACITY',
  OPEN_ENERGY = 'OPEN_ENERGY',
  CLOSED_ENERGY = 'CLOSED_ENERGY',
}